:recycle: Refactored all endpoints into distinct files in directory 'routes' and cleaned up DB connexion into separate file
